requirements.txt
flask
scikit-learn
python-dotenv
requests
